pyproject.toml
[project]
name = "medical-chatbot"
version = "0.1.0"
description = "Medical Chatbot GenAI project using Retrieval Augmented Generation (RAG) with LangChain, Pinecone, HuggingFace and Groq."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-experimental>=0.3.4",
    "langchain-groq>=0.3.8",
    "langchain-openai>=0.3.33",
    "langchain-pinecone>=0.2.12",
    "pinecone[grpc]>=7.3.0",
    "pypdf>=6.1.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.1",
    "streamlit>=1.50.0",
]
